# dist/api-docs.json
{
  "swagger": "2.0",
  "basePath": "/",
  "tags": [
    {
      "name": "Ranking",
      "description": "Veure usuaris ordenats"
    },
    {
      "name": "PlayerAlias",
      "description": "Administració d'usuaris"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/ranking": {
      "get": {
        "tags": [
          "Ranking"
        ],
        "description": "Retorna el llistat d'usuaris totals",
        "operationId": "getRanking",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "": {
            "schema": {
              "properties": {
                "namebreplayers": {
                  "type": "int",
                  "example": "1"
                },
                "players": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/user"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/players/{alias}": {
      "get": {
        "tags": [
          "PlayerAlias"
        ],
        "description": "Retorna informació de l'usuari introduït",
        "operationId": "findByAlias",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Jugador trobat",
            "schema": {
              "$ref": "#/definitions/response200", 
              "properties": {               
                "jugador": {
                  "type": "object",
                  "$ref": "#/definitions/user"
                }
              }
            }
          },
          "504": {
            "description": "Jugador no trobat",
            "schema": {
              "$ref": "#/definitions/response504"
            }
          }
        }
      },
      "post": {
        "tags": [
          "PlayerAlias"
        ],
        "description": "Afegeix un nou jugador",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Atributs del nou usuari",
            "required": true,
            "schema": {
              "$ref": "#/definitions/newuser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Jugador creat correctament",
            "schema": {
              "$ref": "#/definitions/response201"
            }
          },
          "502": {
            "description": "Falta introduïr parámetres correctes",
            "schema": {
              "$ref": "#/definitions/response502"
            }
          },          
          "503": {
            "description": "El jugador ja existeix",
            "schema": {
              "$ref": "#/definitions/response503"
            }
          }
        }
      },
      "put": {
        "tags": [
          "PlayerAlias"
        ],
        "description": "Actualitza un jugador ja existent",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Atributs del nou usuari",
            "required": true,
            "schema": {
              "$ref": "#/definitions/newuser"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Jugador actualitzat correctament",
            "schema": {
              "$ref": "#/definitions/response202"
            }
          },
          "502": {
            "description": "Falta introduïr parámetres correctes",
            "schema": {
              "$ref": "#/definitions/response502"
            }
          },
          "504": {
            "description": "Jugador no trobat",
            "schema": {
              "$ref": "#/definitions/response504"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "position": {
          "type": "String",
          "example": "0"
        },
        "alias": {
          "type": "String",
          "example": "euser"
        },
        "name": {
          "type": "String",
          "example": "example"
        },
        "surname": {
          "type": "String",
          "example": "user"
        },
        "score": {
          "type": "int",
          "example": "0"
        },
        "created": {
          "type": "date-time",
          "example": "2020-01-01T00:00:00.377Z"
        }
      }
    },
    "newuser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "String",
          "example": "example"
        },
        "surname": {
          "type": "String",
          "example": "user"
        },
        "score": {
          "type": "int",
          "example": "10"
        }
      }
    },
    "response200": {
      "properties": {
        "Code": {
          "required": true,
          "type": "int",
          "example": "200"
        },
        "Error": {
          "required": true,
          "type": "String",
          "example": "false"
        },
        "Message": {
          "required": true,
          "type": "String",
          "example": "Player Exists"
        }
      }
    },
    "response201": {
      "properties": {
        "Code": {
          "required": true,
          "type": "int",
          "example": "201"
        },
        "Error": {
          "required": true,
          "type": "String",
          "example": "false"
        },
        "Message": {
          "required": true,
          "type": "String",
          "example": "Player Correctly Created"
        }
      }
    },
    "response202": {
      "properties": {
        "Code": {
          "required": true,
          "type": "int",
          "example": "202"
        },
        "Error": {
          "required": true,
          "type": "String",
          "example": "false"
        },
        "Message": {
          "required": true,
          "type": "String",
          "example": "Player Correctly Updated"
        }
      }
    },
    "response502": {
      "properties": {
        "code": {
          "type": "int",
          "example": "502"
        },
        "error": {
          "type": "string",
          "example": "true"
        },
        "message": {
          "type": "int",
          "example": "The field: name, surname, score are mandatories (the score value has to be >0)"
        }
      }
    },
    "response503": {
      "properties": {
        "code": {
          "type": "int",
          "example": "503"
        },
        "error": {
          "type": "string",
          "example": "true"
        },
        "message": {
          "type": "int",
          "example": "Error: Player Exists"
        }
      }
    },
    "response504": {
      "properties": {
        "code": {
          "type": "int",
          "example": "504"
        },
        "error": {
          "type": "String",
          "example": "true"
        },
        "message": {
          "type": "String",
          "example": "Error: Player not found"
        }
      }
    }
  }
}